Conflicts:
	app/Http/Controllers/SchoolsController.php
	resources/views/course/index.blade.php

+++ app/Http/Controllers/SchoolsController.php
@@ -30,16 +30,55 @@ class SchoolsController extends Controller {
     }
 
     /**
+     * Limit the result set by specified search.
+     * TODO: Refactor this method to a repository for modularity.
+     *
+     * @param Request $request
+     */
+    public function search(Request $request)
+    {
+        $query = Course::active();
+        // Check categories
+        $categories_checked = $request->get('categories');
+        if ( $categories_checked ) {
+            // All courses whose categories match the specified ones.
+            $query->whereHas('categories', function ($q) use ($categories_checked) {
+                $q->whereIn('text', $categories_checked);
+            });
+        }
+        // Check query string
+        if ( $request->get('query_string') ) {
+            $query->where('name', 'LIKE', '%' . $request->get('query_string') . '%');
+        }
+
+        return $query->get();
+    }
+
+    /**
      * Display a listing of the resource.
      *
+     * @param Request $request
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        $schools = School::all();
+        // If search params have been specified, filter the result set
+        if ( sizeof($request->input()) > 0 ) {
+            $schools = $this->search($request);
+        } else {
+            // Else display all
+            $schools = School::all();
+        }
+
         $categories = Category::all();
 
-        return view('school.index', ['schools' => $schools, 'categories' => $categories]);
+        // Flash old input to repopulate on search
+        $request->flash();
+
+        return view('school.index', [
+            'schools'    => $schools,
+            'categories' => $categories
+        ]);
 
     }
 
@@ -85,6 +124,8 @@ class SchoolsController extends Controller {
      */
     public function edit(School $school)
     {
+        $this->authorize('update', $school);
+
         return view('school.edit', ['school' => $school]);
     }
 
@@ -117,9 +158,11 @@ class SchoolsController extends Controller {
      */
     public function removePhoto(Request $request)
     {
-        try {
+        try
+        {
             Photo::destroy($request->input('id'));
-        } catch ( Exception $e ) {
+        } catch (Exception $e)
+        {
             return "Unable to remove photo: " . $request->input('id');
         }
 
@@ -135,6 +178,8 @@ class SchoolsController extends Controller {
      */
     public function update(SchoolRequest $request, School $school)
     {
+        $this->authorize('update', $school);
+
         $school->update($request->all());
 
         return redirect()->action('SchoolsController@show', ['school' => $school]);
@@ -148,6 +193,8 @@ class SchoolsController extends Controller {
      */
     public function destroy(School $school)
     {
+        $this->authorize('update', $school);
+
         $school->delete();
 
         return redirect()->action('SchoolsController@index');
